curlhttpsrc|curlhttpsink: Support PINNEDPUBLICKEY
Add support for curls CURLOPT_PINNEDPUBLICKEY.

From curl documentation:
Pass a pointer to a null-terminated string as parameter. The string can be the file name of your pinned public key. The file format expected is "PEM" or "DER". The string can also be any number of base64 encoded sha256 hashes preceded by "sha256//" and separated by ";"

When negotiating a TLS or SSL connection, the server sends a certificate indicating its identity. A public key is extracted from this certificate and if it does not exactly match the public key provided to this option, curl will abort the connection before sending or receiving any data.