mxf: mux/demux distinguish random access points from keyframes
`mxfdemux` is capable of building an index on the fly when it is not included
in the original stream. However, it made no distinction whether key frames were
part of an open or closed gop. When a seek occurred, the first preceding key
frame was selected even if it was part of an open gop, which could lead to
decoded frames being dropped due to backward prediction to an anterior keyframe.
In case of a seek in paused state, preroll would fail.

The "MXFIndexEntry" includes flags which are capable of representing a keyframe
not being a "Random Access Point". Additional flags can also encode whether the
frame uses backward and or forward prediction. The demuxer can make use of
these precisions in case of a seek to decide whether it needs to go back to
latest or penultimate keyframe for the seeked frame to be decoded successfully.

This commit uses these flags throughout the relevant code:

* mxfdemux: improve the `find_*_offset` algorithm to identify the frame
  dependencies while searching back the keyframe.
* mxf codec handlers: the handlers used to (un)set the DELTA_UNIT flag on the
  frame buffer to indicate whether it was a keyframe. Codec handlers are now
  provided with a dedicate `struct` which can hold the MXFIndexEntry flags,
  and other precisions which can be useful in case of an index-less stream.
  The mpeg codec handler was updated to fill this `struct` with the relevant
  data.
* mxfmux: use the codec handlers to build a more accurate index.