ges: remove SUPRESS_UNUSED_WARNING
It seems to be a mechanism to avoid compiler warning,
that appears on clang compiling code with G_DECLARE_FINAL_TYPE,
because inside it was defining some inline functions that could
never be used.
Starting from 2.63 glib have this warning fixed.

It was done in this commits:
https://gitlab.gnome.org/GNOME/glib/-/commit/d58be1bda9b270a618f341bf7d9b0dd88702482c
https://gitlab.gnome.org/GNOME/glib/-/commit/90f93a64f26666b955fdafb56acab54e596f4cc5

Also there're already some other pieces of code in GStreamer
that do use G_DECLARE_FINAL_TYPE without supressions.