ci: Check formatting of python files
We already have a git hook which enforces the code style, however
it's not a stnadalone script and it only checks staged files.

This adds an optional argument that can be passed to the hook
script, that will make it look at all the python files in the
repository instead of the staged ones.

Instead of factoring out the codestyle script outside the git hook
add an optional argument to the existing one, since we only going
to use the extra functionality in the CI.

Close #2930