mxf: mux/demux: build temporal offsets for index-less streams
Index-less streams need a temporal offset list to apply pts-dts conversion and
reach the expected frame when a stream contains B-frames.

This commit adds a dedicate index for index-less streams. We don't try to
rebuild the Segment and its `MXFIndexEntry` array as these include more data
which is not available in this context ans also because the index can only be
gradually built. The addition of the new index is quite contained though since
it is used via the function `pts_to_dts_position`, so it makes no difference to
the caller whether the conversion implied on index or the other.

Also fixes a sign issue with `key_frame_offset` in mxfmux. See st381-2:

> Key frame offset is always negative for MPEG-2 streams.