v4l2: make sure source change event is reported
The source change event may not be reported if the format set
by OUTPUT is the same with that is parsed. It will cause decoder
to be waiting for the source change event and hang. So set the
width and height to zero to make sure source change event is reported.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/5396>