webrtcbin: Fix transceiver matching with FEC decoder for PCMU
When using PCMU, the payload type can be 0. When `rtpbin` requests
FEC decoder and passes a payload type of 0, the check on payload
type in `try_match_transceiver_with_fec_decoder` resulted in the
FEC decoder not being set on the transceiver. This then resulted
in code setting pass through on the `rtpulpfecdec` being bypassed
in `_set_internal_rtpbin_element_props_from_stream`. We need to
set pass through on `rtpulpfecdec` as it is broken.

https://gitlab.freedesktop.org/gstreamer/gst-plugins-good/-/issues/581
https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/3212