downloadbuffer: Unlock when receiving non-flushing seeks
The element already unlocks its getrange() when it receives a flush-start, but
that event is usually not received when the seek is operating only on the
segment, without flushing.

This patch should fix deadlocks happening in WebKit when a video element is
looping on a media file hosted on a HTTP server and with preloading enabled.

Fixes #3129