alsasrc: Handle snd_pcm_readi() errors correctly
According to libasound documentation the return values of
snd_pcm_readi() are either the frame count or one of EBADFD, ESTRPIPE,
EPIPE.

Some code reading in libasound as well as the underlying ioctl()
implementation reveals that there are more possible errors than just
these three. However, the original code was wrong as DISCONNECTED is
reported as EBADFD (implicit as invalid status) and ENODEV will never
be returned.

Anyway, except for EPIPE which reports an overrun all other errors are
unrecoverable within the running pipeline.

Signed-off-by: Jochen Henneberg <jh@henneberg-systemdesign.com>