queue: ignores min-threshold-* when serialised event is sent downstream
By ensuring we do not treat the queue as filled if an event is in the tail
in order to avoid draining the queue before min-thresholds are reached.
But we release events immediately if they are at the head of the queue
as its not clear whether they should wait for buffer thresholds to be
met.