ci: Build and run gstreamer on Lava
Build GStreamer for the arm64 device with v4l2codecs support. This build
is uploaded to s3.freedesktop.org so that Lava can retrieve it.

The Lava job creation is based on Mesa CI. It creates a Lava job that
downloads the GStreamer build from the S3 server, setup the correct
PATHs and runs fluster-runner.

When the Lava test finishes, the results are retrieved from the S3
server and shown on the gitlab CI tests results.