aggregator: handle erroneous case of the first buffer before segment
Segment event always should arrive before the first buffer, however if
that happens by accident due to some complex pipelines manipulations,
we don't want the process to crash.
In particular an abort gets raised by the videoaggregator in this case.
To avoid that we propose to safely assume we have received a segment
that starts from the first buffer's pts.