qt: fix segmentation fault in QtGLVideoItem::onSceneGraphInitialized
`QtGLVideoItem::onSceneGraphInitialized` can be invoked by runable in Qt
render thread just after (during?) destruction of `QtGLVideoItem`. Once
`this->priv` is set to `NULL`,  segmentation fault occurs during access
to `this->priv->qt_context` or even in `GST_DEBUG()` call if `qt*:5` is
enabled.

Check if `this->priv` is NULL and short-circuit out of
`QtGLVideoItem::onSceneGraphInitialized` to avoid crash.

Closes #2421