gst-env: Add a workaround for PATH length limits on Windows
Sometimes on Windows we have too many directories with executables and
DLLs, and adding them all to PATH causes the console to behave very
weirdly, likely because PATH has a length limit. This limit can be
increased to 32,767, but that isn't enough to fix the issue on some
systems:

https://learn.microsoft.com/en-us/windows/win32/fileio/maximum-file-path-limitation

This makes the devenv completely unusable for these people on Windows.

Let's work around that by giving people the option of symlinking every
EXE and DLL into builddir/meson-uninstalled/bin and just adding that
to PATH. This needs Windows 10 1703 or newer (released 7 years ago)
and "Developer Mode" to be enabled in the Settings.

Also update the documentation to mention the meson install
alternative, and move to OrderedSet.