info: bring lazy initialization of the categories into API
This improves the usability of the framework, because it
allows to avoid calling GST_DEBUG_CATEGORY_INIT().
By avoiding calling GST_DEBUG_CATEGORY_INIT() we
get rid of the "uninitialized category" error as a type,
don't have to think where to put it, and in the end
it allows us to write less code.
At a first impression it may seem that the lazy initialization
of the category has some micro performance cost, but in fact
it can also be seen the way around: if GST_DEBUG is not set,
the category won't be initialized at all, that is a micro
performance win.