gst-plugins-bad: Add suppression for libsrt pthread singleton
The SRT library uses pthread_create_key and pthread_setspecific to store
the latest error on a thread as a singleton.

At thread exit the specified destructor should run and free that
singleton, failing that some C++ destructors should perform it.

This does not always happen, in our tests the reason is that
pthread_exit is not called on main threads.

In order for our valgrind tests to passed we will add suppressions for
this leak of the singleton error.