gstvalue: Transform an empty array to an empty string
A null-pointer is a valid value for an empty `GValueArray`. An empty
array is a default value for an array property. Since
`_gst_value_transform_g_value_array_string` doesn't check if the array
is empty, one that calls `g_param_spec_get_default_value` and tries to
convert the returned value to a string ends with a segmentation fault.
For example, `libglibmm_generate_extra_defs-2.68.so` does this.

Fixes #2874