macos: Add sckitaudiosrc element
Built with SwiftPM for now, because Meson support for Swift is basically none.
Uses Obj-C to call into Swift (via a bridging header) and the Swift side can directly call into C via a module
that imports Gst headers.

Only an audio element for now. Video will be a bit more complicated, and once it's needed, it will be best to split
the shared parts between both into a base class or so.

Does not allow to specify a window to capture from or anything, just system-wide audio for now. Needs figuring out
how to specify what we want to capture (bundle identifiers?), what to exclude etc.