vabaseenc: refactor rate control property handling
Instead of handling the property in each element, repeating a lot of code, all
is refactored in as a helper in the base encoder class. Thus the property is the
first one to be installed, and it's always installed. If the driver doesn't
provide any rate control, only NONE is exposed.

Fixes #3376