qtdemux: Use the smallest negative composition offset of the different fragments
As the minimal negative composition offset is not constant across fragments, we can have this
case:

-Fragment 1: has 100 samples
The minimal negative composition offset is equal to -2.
The last sample (buffer) of this fragment has the DTS = 100 and the CT (composition time offset) = 0,
so its PTS= 100 + 0 + 2 = 102

-Fragment 2: has 100 samples
The minimal negative composition offset is equal to -1.
The first sample has the DTS = 101 and the CT (composition time offset) = 0,
so its PTS= 101 + 0 + 1 = 102

Thus, the last sample of the first fragment has the same PTS as the first sample of the
second fragment, that causes an issue in the video decoder.

With this commit the minimal negative composition offset of the second fragment is the
smallest negative composition offset of the both fragments so is equal to -2, so
the PTS of the first sample is equal to 101 + 0 + 2 = 103.

This solution ensures that the PTS increment, but can cause a hole in the PTS when the minimal
negative composition shift decreases (e.g. from -1 to -2) between two fragments.

ISO/IEC 14496-12:2022 specification (mp4/ISO BMFF) is a bit fuzzy about these aspects and does not help
from our point of view, 'cslg' box presence with compositionToDTSShift should be mandatory for fragmented mp4
when negative values are used and the minimal value of sample_composition_time_offset is not constant across fragments.

Also-by:philippe.grave@softathome.com